first pass at splitting up server.js
